+++ src/main/java/org/jufe/anmeldetool/entity/BaseEntity.java
@@ -14,4 +14,16 @@ public class BaseEntity {
 
     protected BaseEntity() {}
 
+    public BaseEntity(UUID id) {
+        this.id = id;
+    }
+
+    public UUID getId() {
+        return id;
+    }
+
+    protected void setId() {
+
+    }
+
 }
