+++ src/main/resources/application.properties
@@ -1,6 +1,12 @@
-debug=true
+logger.0.level=info
+logging.level.root=info
+logging.level.org.hibernate=trace
+logging.level.org.springframework.web=debug
 
-spring.datasource.url=jdbc:h2:file:./db
+debug=true
+spring.l
+
+spring.datasource.url=jdbc:h2:mem:./db
 spring.datasource.username=sa
 spring.datasource.password=
 spring.datasource.driverClassName=org.h2.Driver
@@ -8,7 +14,7 @@ spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
 spring.h2.console.path=/h2-console
 spring.h2.console.settings.trace=false
 spring.h2.console.settings.web-allow-others=false
-spring.jpa.hibernate.ddl-auto=create
+spring.jpa.hibernate.ddl-auto=create-drop
 
 # Beispielproperties. Nur für den Fall, dass wir irgendwann man was konfigurieren müssen
 anmeldetool.host
